Transliteration of Exponential Reservoir Decay
The yammer / codahale metrics library implements a nice sample method
that provides for expoential decay with _tiny_ amounts of memory.

This is based on the Comode et al paper "Forward Decay: A practical Time
"Decay Model for Streaming Streams"
(http://dimacs.rutgers.edu/~graham/pubs/papers/fwddecay.pdf)

This hopefully provides a transliteration of the java code to scales,
done with the java code in one hand, whilst glancing at the Comode paper
in the other.

Acked-by: Zac Propersi <logicalbomb@github.com>